+++ packages/cedar/src/Chart.ts
@@ -1,7 +1,7 @@
+import { queryFeatures } from '@esri/arcgis-rest-feature-service'
 import { cedarAmCharts } from '@esri/cedar-amcharts'
 import { flattenFeatureSet, flattenFeatureSets } from './flatten/flatten'
-import { getData } from './query/query'
-import { createQueryParams, getQueryUrl } from './query/url'
+import { createQueryParams } from './query/url'
 
 function clone(json) {
   return JSON.parse(JSON.stringify(json))
@@ -169,9 +169,11 @@ export default class Chart {
         if (dataset.url) {
           // TODO: make name required on datasets, or required if > 1 dataset?
           names.push(dataset.name || `dataset${i}`)
-          const queryUrl = getQueryUrl(dataset)
           const queryParams = createQueryParams(dataset.query)
-          requests.push(getData(queryUrl, queryParams))
+          requests.push(queryFeatures({
+            url: dataset.url,
+            params: queryParams
+          }))
         }
       })
     }